implements InstructionBox as uikit component
